feat: implement pass logic for application tests
