+++ __tests__/OutputTest.js
@@ -1,3 +1,4 @@
+import Lotto from "../src/Lotto";
 import {
   printProfitRate,
   printPurchaseCount,
@@ -30,8 +31,8 @@ describe("출력 처리 테스트", () => {
     const logSpy = getLogSpy();
 
     printPurchasedLottoNumbers([
-      [8, 21, 23, 41, 42, 43],
-      [3, 5, 11, 16, 32, 38],
+      new Lotto([8, 21, 23, 41, 42, 43]),
+      new Lotto([3, 5, 11, 16, 32, 38]),
     ]);
 
     const logs = ["[8, 21, 23, 41, 42, 43]", "[3, 5, 11, 16, 32, 38]"];
@@ -44,16 +45,10 @@ describe("출력 처리 테스트", () => {
   test("구매한 로또 개수에 따른 로또 번호 출력(예외 케이스)", () => {
     expect(() =>
       printPurchasedLottoNumbers(
-        [8, 21, 23, 41, 42, 43],
-        [3, 5, 11, 16, 32, 38],
+        new Lotto([8, 21, 23, 41, 42, 43]),
+        new Lotto([3, 5, 11, 16, 32, 38]),
       ),
     ).toThrow("[ERROR]");
-  });
-
-  test("구매한 로또 개수에 따른 로또 번호 출력(예외 케이스)", () => {
-    expect(() => printPurchasedLottoNumbers([3, 5, 11, 16, 32, 38])).toThrow(
-      "[ERROR]",
-    );
   });
 
   test("구매한 로또 개수에 따른 로또 번호 출력(예외 케이스)", () => {
